refactored: split index service in other files
