+++ src/types/index.ts
@@ -1,0 +1,5 @@
+export interface ChatMessage {
+    role: string;
+    content: string;
+  }
+
